pyproject.toml
[tool.poetry]
name = "uraf"
version = "0.2.0"
description = "Unified Reasoning and Aggregation Framework (URAF) for evaluating LLMs with advanced reasoning capabilities."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "uraf" }]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"                  # Data analysis and manipulation
requests = "^2.32.3"               # LLM API interaction (e.g., LM Studio, OpenAI)
openai = "^1.61.0"                 # OpenAI API client for LLMs
langchain = "^0.3.17"              # Advanced prompt chaining and reasoning
tqdm = "^4.64.1"                   # Progress bars for batch evaluations
numpy = "^1.23.0"                  # Numerical computations
flashtext = "^2.7"                 # Ultra-fast keyword extraction
keybert = "^0.7.0"                 # BERT-based keyphrase extraction
sentence-transformers = "^3.4.1"   # Embedding-based similarity for evaluations
evaluate = "^0.4.3"                # NLP evaluation metrics (e.g., BLEU, ROUGE, METEOR)
datasets = "^3.2.0"                # Hugging Face datasets for benchmarks
bertopic = "^0.15.0"               # Topic modeling with BERT
umap-learn = "^0.5.3"              # Dimensionality reduction for topic modeling
hdbscan = "^0.8.33"                # Clustering for topic modeling
scikit-learn = "^1.3.0"            # Required for BERTopic
rouge-score = "^0.1.2"             
loguru = "^0.7.3"
asyncio = "^3.4.3"
aiohttp = "^3.11.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"                    # Unit testing
black = "^23.0"                    # Code formatting
flake8 = "^6.0"                    # Linting for style checks
mypy = "^1.2.0"                    # Static type-checking
isort = "^5.12.0"                  # Organize imports
pytest-mock = "^3.10.0"            # Mocking for tests
ipython = "^8.0"                   # Interactive Python for quick experiments
python-dotenv = "^1.0.1"           # To use evaluate-metric/rouge

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
